pip-requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
classes==0.1.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
idna==2.9
lxml==4.5.0
pycparser==2.20
PyMySQL==0.9.3
python-docx==0.8.10
python-dotenv==0.12.0
python-telegram-bot==12.7
pytz==2020.1
requests==2.23.0
schedule==0.6.0
six==1.14.0
SQLAlchemy==1.3.16
tornado==6.0.4
typing-extensions==3.7.4.2
urllib3==1.25.8
